+++ openhands/utils/embeddings.py
